+++ backend/core/src/listings/dto/listing.dto.ts
@@ -691,4 +691,12 @@ export class ListingFilterParams extends BaseFilter {
     required: false,
   })
   status?: ListingStatus
+
+  @Expose()
+  @ApiProperty({
+    type: String,
+    example: "Fox Creek",
+    required: false,
+  })
+  neighborhood?: string
 }