+++ backend/core/src/shared/dto/filter.dto.ts
@@ -7,6 +7,12 @@ export enum Compare {
   "<>" = "<>",
 }
 
+export enum Filters {
+  "status" = "STATUS",
+  "name" = "NAME",
+  "neighborhood" = "NEIGHBORHOOD",
+}
+
 export class BaseFilter {
   @Expose()
   @ApiProperty({